requirements.txt
alabaster==0.7.12
alembic==1.4.3
argh==0.26.2
argon2-cffi==20.1.0
asn1crypto==1.3.0
astroid==2.4.2
astropy==4.1
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
cryptography==3.1
culqi==1.0.0
culqipy==0.2.6
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dnspython==2.0.0
docutils==0.15.2
email-validator==1.1.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
fpdf==1.7.2
future==0.18.2
gunicorn==20.0.4
helpdev==0.7.1
idna==2.8
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.3.0
intervaltree==3.0.2
ipinfo==3.0.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==5.5.3
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.3
jplephem==2.13
json5==0.9.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab-server==1.0.6
keyring==20.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
MyQR==2.3.1
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.1.3
numpy==1.18.1
numpydoc==0.9.2
opencv-python==4.4.0.46
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psutil==5.6.7
psycopg2==2.8.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.6.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.6
PySocks==1.7.1
pytesseract==0.3.2
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3
pyzbar==0.1.8
pyzmq==18.1.0
QDarkStyle==2.8
qrcode==6.1
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
rope==0.14.0
Send2Trash==1.5.0
sgp4==2.2
six==1.14.0
skyfield==1.31
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spotipy==2.12.0
spyder-kernels==1.8.1
SQLAlchemy==1.3.16
svgwrite==1.4
terminado==0.8.3
testpath==0.4.4
timezonefinder==4.4.0
toml==0.10.1
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
wtforms-html5==0.1.3
xlrd==1.2.0
zipp==0.6.0
